refactor: format code and e2e
