+++ e2e/fixtures/originImage/PostProcess_postProcess-HDR-bloom-ACES.jpg
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:970180594182c8ae8926539c6ff8a43b2b0446ed244667d2f770724fbc6f3548
+size 395920
