+++ e2e/fixtures/originImage/PostProcess_postProcess-HDR-bloom-neutral.jpg
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:42bd43b5d7276f070c2a530cdbd3222be9208d72a99eaf758b381128d02a8bd8
+size 395180
