+++ e2e/fixtures/originImage/PostProcess_postProcess-LDR-bloom-neutral.jpg
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:b91b0f1aae505040173e26910095412f52c7b8028847d835ebef8db89c8c0cec
+size 342057
