--- e2e/fixtures/originImage/PostProcess_postProcess-bloom-tonemap.jpg
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:2d31cc04e4214967d470c2e9b0ca85566dfa61481cae19563c91f125fe8bb840
-size 314694
